+++ tests/flow/test_flow.py
@@ -124,7 +124,8 @@ class TestInferencePipeline(unittest.TestCase):
                         input_size=chunk.shape,
                         output_patch_overlap=self.patch_overlap,
                         framework='identity',
-                        batch_size=5) as inferencer:
+                        batch_size=5,
+                        dtype='float32') as inferencer:
             print(inferencer.compute_device)
             chunk = inferencer(chunk)
         print('after inference: {}'.format(chunk.slices))
@@ -144,9 +145,9 @@ class TestInferencePipeline(unittest.TestCase):
         print('save to output volume...')
         save_operator = SaveOperator(self.output_volume_path,
                                      self.mip,
-                                     upload_log=False,
+                                     upload_log=True,
                                      create_thumbnail=True)
-        save_operator(chunk)
+        save_operator(chunk, log={'timer': {'save': 34}})
         print('after saving: {}'.format(chunk.slices))
 
         # evaluate the output